Config2PluginxxxxMerge: Put the pane in the right place
When Sys/Font/Boot merge are run in a mode < 1024 pixels tall, the act of creating the toolbox objects for the main window plus pane is subject to the Wimp clipping the windows to the top of the screen they were created in in the Res file.
So, if used in a less tall mode (say 800x600) the pane is wrongly offset in the agenda view.
Edit the Res files to be -512, -1024 OSU down in the bottom left corner to avoid this.

Version 2.09. Tagged as 'xxxxMerge-2_09'