Draw: Static analysis fixups
Line 114:
Remove the potential divide by zero by considering that the only way to get xspace=0 is when yspace <= 1, in which case it's a unity grid.
Line 295:
Add some brackets to make the expression unambiguous. Fortunately with Norcroft ((thing & 1) == 1) results in the same output as (thing & (1 == 1)) in this situation.
From a tip off in https://www.riscosopen.org/forum/forums/4/topics/3990

Version 1.29. Tagged as 'Draw-1_29'