Pinboard: Emit boot relative Pin/AddTinyDir commands where possible
When saving the desktop settings, use the same substring matching logic that !Configure's boot addapps/lookat/run uses, and replace the base of the name with Boot: or Boot:^.
This means that if the boot drive is renamed the pinboard still starts up with the same items on it.
Ditch support for Message_SaveDesktop - the code's been a NOP since Ursula (circa 1998).
Tested on a Raspberry Pi, renaming the SD card.

Version 1.01. Tagged as 'Pinboard-1_01'