ScrnSetup: Make monitor type into a display field + popup menu
Was previously a StringSet, but these are limited to only one level of menu tree.
Turned off "Deliver event before showing" for the popups in the main window, since nothing registered events against them.
No longer spends an eternity rebuilding the monitors menu when adjust clicking on 'Cancel', because the read_choices() function also used to scan the MDFs to make the menu. Now it really does just read the choices back in.
Correction to interactive help for the "Resolution" menu - the guidance about square and rectangular pixels was back to front.

Version 0.35. Tagged as 'ScrnSetup-0_35'