ScrnSetup: Group manufacturers together in the type menu
The plugin scans all the subdirectories of BootResources:Configure.Monitors but then presents the findings unsorted in a huge menu. Now there are lots of MDFs the menu is so big it has a scrollbox too.
To simplify this, and make finding the right one easier, collect up the monitors by manufacturer (inferred from name of the topmost directory) into submenus.
Tested on a Risc PC with a mix of complexities of subdirectories from a default !Boot.

Version 0.36. Tagged as 'ScrnSetup-0_36'