InetRes: Add Tftp to the disc based !Internet commands
A TFTP client is described in the TCP/IP Protocol Suite manual but seems to have been lost somewhere along the way, perhaps when the Internet 5 work was done? TFTP is often used to load OS images or similar boot time stuff as it's a very simple (and insecure) protocol.
Port a contemporary client from FreeBSD (circa March 2000) with a few changes to allow it to run on RISC OS.
Makefile:
  Remove the BareMinUtils option. Nothing uses it, and it wouldn't work anyway since it's missing !Run.
  Add Tftp

Tested against OpenTFTP server 1.64 running on Windows, GETting and PUTing a number of files.

Version 5.58. Tagged as 'Internet-5_58'