SystemDevs: Add file-like-devices for zero: random: urandom:
Add rough parallels to Unix /dev/zero /dev/[u]random alongside existing null.
To avoid consuming lots of FS numbers we lump these all together as "SourceFS" and use special fields to select which is wanted. Other sources can be trivially added by extending the special field table.
Currently [u]random are stateless, in that there is one global seed for all open random sources, and the PRBS is based on the algorithm for rand() in the ISO9899 C standard - in future more sources of entropy could be mashed in.
To use, simply OPENIN"zero:" or "random:" "urandom:" and BGET or OS_GBPB as much data as you want - EOF is never reached.
Misc other changes:
  Get 'standalone' passed in from the Makefile, rather than work out 'standalonemessages' here.
  Remove unused "serial:" support switch.
  Replace a few magic numbers with values from Hdr.

Version 1.33. Tagged as 'SystemDevs-1_33'