FilerAct: Fix for NULL pointer dereference skipping/retrying the last file
If an error occurs during an operation (for example, during a copy move where the source is on a read only medium) the status window changes to an error window with "Abort/Skip/Retry". The skip and retry options call read_next_node_parameters() to peek the next queued item, but would attempt to OS_File 5 of a NULL filename if the error happens when there are no more.
Check the next_nodename isn't NULL to avoid an abort in OS_GSInit with zero page protection enabled.

Version 0.62. Tagged as 'FilerAct-0_62'