NetFS: Audit of signed pointer compares
Fairly thorough combing of things holding addresses with condition codes LT GE LE GT.
The compare in Configure.s was missed when the exact same bug was fixed in revision 4.3 of this file.
Given a reasonable test with FilerAction browsing around a standard disc image exported as an FS via Level 4.

Version 5.99. Tagged as 'NetFS-5_99'