SoundCtrl: MixVolume parsing correction
SoundCtrl is supposed to accept "0" to mean the default mixer where not written out longhand, for example as written by earlier versions of the !SndSetup configure plugin from RISC OS 5.00 up to and including SndSetup-2_07. However, the match for "0" never happened and the MixVolume command was subsequently ignored.

Version 1.06. Tagged as 'SoundCtrl-1_06'