Omni: Add support for drag-to-iconbar for all of Omni's clients
The bones of support for this were already present, but disabled by the 'EXPERT' switch.
Since the other filers have supported the operation since Ursula, enable & refactor the code to use the same protocol the other filers use.
Allow a non root destinate to be used if the variable 'FSTitle + Filer$DefaultDir' is set.
Remove dead UserDrag code.
Tested with LanManFS and (Omni) NetFiler.

Version 2.24. Tagged as 'Omni-2_24'