SDCMOS: Drop StoreInHAL switch
No longer support or maintain the code which opened riscos.img and wrote values into it on the Pi.
However, carry over the logic which attempts to restore the time from the timestamp of the CMOS file from last shutdown so that the time (on a clockless Pi) is a better guess than the Unix epoch until the network time can be read.

Tested on a Pi 3.

Version 0.17. Tagged as 'SDCMOS-0_17'