IIC: Error number alignment & minor tidying
The trunk version of IIC uses &20301 as "IIC in use" by defining it locally at the top of the source, whereas the RISC OS 5 kernel (using the HAL branch of IIC) uses &20302. Since the trunk version with reentrancy checking was only ever shipped in STBs, take the allocation in Hdr:NewErrors as gospel and swap to using that.
Other minor tidying:
* Use Push/Pull instead of STM/LDM
* Beef up some comments
* Use Unix style filenames in the Makefile
Built, but not tested.

Version 0.18, 4.5.2.9. Tagged as 'IIC-0_18-4_5_2_9'