FileCore: Fix for nonsense sector offsets (and abort) with *Map of old map disc
For new map discs NextFs returns the address of the map zone in R10, which was being used to lookup entries in the disc record. However, for old map discs R10 is a multiple-of-3 offset into the free space map. This was aborting on the 2nd call due to an unaligned LDR at line 3205.
Even on platforms that permit unaligned loads the printed results were wrong because the disc address for an old map disc is in bytes but the value printed out was shifted up by Log2SectorSize. A DD floppy disc could have free space fragments starting at 7M - nonsense.
Fixed by checking the DiscFlags and only reading from the map zone/sector shifting for new map.
Tested with an ADFS-L floppy disc.

Version 3.71. Tagged as 'FileCore-3_71'