OmniLanManFS: Resolved issue where RISC OS name '1 2/3' gets wrongly translated.
Detail:
The translate routine that converts contentious characters in  a RiscOS
file name to equivalent DOS ones was being invoked a second time. Normally
this wouldn't be critical, but a / in riscos was converted to the . in DOS,
correctly , then converted back again. This error mode with more than one
contentious character in (particularly) a directory name had recently
come to light.
Admin:
Tested by building a RiscOS rom image on a pc based LanManFS mount without
errors.

Version 2.60. Tagged as 'LanManFS-2_60'