FontManager: Fix for failure to plot bitmap fonts when the font cache is above 256MB
The bitmapped font paintmagnified routine uses bit indexes, which needs an effective 35 bit address (for a 4bpp IntMetrics+x90y45 font).
In preparation for Medusa (>=256MB machine) the top 3 bits are kept in paint_inptr and the bottom 32 bits in the inptr register (R4).
However, a mixup meant the top 3 bits from paint_inptr were added before an extra 2 bit shift, so the font cache address becomes nonsense, leading to an abort.

Fonts02.s: Reorder the calculation.

Tested with Porterhouse bitmap font, with the font cache at &37000000 (ie. one of the top 3 bits are set), using it as the desktop font, in !Edit, printing to a PostScript and bit image printer, and in !TechWriter.

Version 3.78. Tagged as 'Manager-3_78'