SprExtend: Fix to dpcm calculation leading to divide by zero in !Draw
For JPEGs specified in dots-per-cm some dodgy looking bytewise fractions and missing brackets meant the x and y densities always came out as zero, leading to an abort in !Draw when it asks JPEG_[File]Info for this to convert to internal units.
Correct the calculation and split it onto 2 lines for readability.

Tested with a sample JPEG at 37dpcm (~93dpi).

Version 1.79. Tagged as 'SprExtend-1_79'