Configure: Don't postmortem when merely run at the CLI
FontMerge is intended to be a Wimp task, but only installs a signal handler for SIGOSERROR after calling Wimp_Initialise. Therefore, when invoked from the command line the "Window Manager is in use" error ends up with the default signal handler in the CLib, which postmortems.
Install a simple non-wimp error handler.
Also, don't try to clone !Sprites23 when adding supplementary (numbered) !Fonts directories when dir is full, these haven't been provided since Ursula.
Version manually bumped to 0.03, since FontMerge is a sub component of Configure.

Version 1.92. Tagged as 'Configure-1_92'