ADFS4: Run time detect the sector size
Read the sector size from ATA Identify, and keep it in the drive structure. This allows a mix of sector sizes to be used across discs.
Search & replace LOG2_SECTOR_SIZE and SECTOR_SIZE with the value read from the drive.

Tested on Titanium with an assortment of manufacturer drives, capacities, and sector sizes.

Version 4.03. Tagged as 'ADFS-4_03'