DebugTools: Resolve two differing canonicals
Disc based canonical (supplied with the DDE) takes 1 arg, the variable to expand. The version in this module also requires a path name, doing the assignment too.
Take the disc version to be canonical (!) since that one's in the wild, and remove the 2nd arg from DebugTools.
Docs and test case updated accordingly.
Also:
* Use Unix paths (thing.s not s.thing).
* Makefile recreated from fragments.

Version 1.20. Tagged as 'DebugTools-1_20'