XHCIDriver: Use RTSupport for both stalls and aborts
Implement the 'todo' note to use RTSupport instead of a callback. This means if the OS is stuck somewhere in SVC mode and no callbacks are happening the pipe is aborted, plus the state is held in the xfer struct rather than a single shared variable.

Version 0.18. Tagged as 'XHCIDriver-0_18'