HForm: Increase support for formatting up to idlen of 21
Check which version of FileCore we have, and use that to pick an appropriate idlen (in a similar manner to idlen of 19 or 15 was chosen for BigDisc support).
As idlen is increased so to are the number of zones, and so to is the size of the map to hold those zones inside HForm. Rather than ballooning the map DIM again, reduce it back to its pre-Ursula size of &40000 and instead write the new map out as it is generated.
Fix for longstanding syntax error in PROCWriteOldFsMap, an errant 'p' had been added.
Changed NukeDOS, and its name, to use the existing Buf% rather than leaking a DIM of its own.

Tested with a range of 512B sector drives manually picking LFAUs to get idlen=17 to idlen=21 inclusive, then cross checking with DiscKnight 1.52.

Version 2.74. Tagged as 'HForm-2_74'