WinEdit: Fix for NULL pointer dereference opening icon properties
WinEdit keeps a track of the previous caret position, so that as windows (within itself) are closed the caret can be put back in the previous window.
However, when there's no previous caret owner (ie. it's the first properties window opened) lastwindow is NULL, and an attempt to read from a window structure in zero page is made.

props.c: copy the NULL check from a couple of lines higher up in the same function.
Makefile: put the objects into OBJS so a debug build can be made.

Version 1.12. Tagged as 'WinEdit-1_12'