ABRelease: Update !SysMerge script for BASIC-1_72
To handle the situation where the disc based BASIC V has *not* been loaded, but the disc based BASIC VI has, add a dummy obey file to the Library so that it can start *BASICFPA instead.
At some future point this can be removed when the disc based BASIC combines all 3 floating point types in one.

Version 0.83. Tagged as 'ABRelease-0_83'