DOSFS: Allow mounting discs with 1 letter names, and 2x buffer overrun fixes
DOSFS would prefer to use the disc title from the volume name entry in the root directory, even if this reduced to 1 letter, which would then lead to "Disc drive not known" errors. Since the disc couldn't be mounted it couldn't be renamed either, requiring a trip to a PC to do so.
Now, if the cleaned up name is < 2 characters long, ignore it and use the one based on the volumeid, and if that doesn't exist, use the default one ("Untitled") from the Messages file.
In OpsFunc.c used memset() in place of a byte set loop.

Ref https://www.riscosopen.org/forum/forums/4/topics/9503
DOSFS.c, line 621: remove the excess NULL so the sprintf fits into the 10 byte buffer, not 11.
DOSFS.c, line 1556: restore the original intent for *CopyBoot in light of the structure/union reorg DOSshape.h revision 4.7 in DOSFS-0_79. Previously the boot code was copied in its entirity as an opaque blob, but the loop was left unchanged, overrunning the sector buffer.

A useful example of the extent of the boot code is given in
  http://thestarman.pcministry.com/asm/mbr/DOS50FDB.htm
which shows the code uses up every last byte of the sector.

Tested with a disc called "R", which can now be mounted. A floppy with no name and no volume id, and one with no name and a volume id.

Version 1.10. Tagged as 'DOSFS-1_10'