Kernel: Adapt A15 KnownCPUFlags to account for LDREX and friends
When SyncLib-0_04 is used in conjunction with SDIODriver, the exclusive access instructions (instead of SWP) abort when the cache is off, causing the machine to hang if *Cache Off is used.
See A15 TRM section 6.4.5 for the logic.
Copy the A53's known CPU flags to prevent the D cache being turned off.

Version 5.97. Tagged as 'Kernel-5_97'