Window: Button_GetFlags method fixes
The gadget copy of the Wimp flags was not being correctly updated, should be
  newflags = (oldflags BIC clear_word) ^ eor_word
per analysis in ticket #442.
In passing, it was also noted that the eor_word was not being masked, meaning it would be possible to toggle the indirect icon flag - breaking the rules set out on the User Interface Toolbox manual for that gadget type.

button.c: Fix flags copy, and mask correctly. Trim some superfluous breaks.
draggable.c: Fix out of bounds array access, found by static analysis.

Version 1.79. Tagged as 'Window-1_79'