XHCIDriver: On stall cancel the transfer's timeout too
When an endpoint stalls, make sure the xfer's timeout is cancelled before setting up the background clear operation, otherwise the stall is cleared and a few seconds later an attempt is made to work with the (stale, or possibly even reused) xfer struct in xhci_timeout().
Merged from NetBSD revision 1.28.2.74.

Version 0.26. Tagged as 'XHCIDriver-0_26'