HAL_Titanium: Add missing module to key scan dependencies
In some circumstances XHCIDriver uses an RTSupport thread, but RTSupport was missed off the list of modules. Add it.
Fixes problem with PixArt (093A/2510) OEM mouse getting stuck at power on (it stalls SetIdle transactions), but which worked if plugged in after power on when RTSupport is loaded.

Version 0.06. Tagged as 'Titanium-0_06'