GPIO: Fix for only returning the first and last GPIO controller's enumeration
The loop to search for the next controller to return for GPIO_Info would only find the first and last controller and skipped any intermediate ones.
For the Raspberry Pi, with 2, this went unnoticed.
Also, the inner loop would read off the end of the gpio_device array by 1 entry.

Tested on a Raspberry Pi (results unchanged) and OMAP3 (6 controllers).

Version 1.00, 1.11.2.5. Tagged as 'GPIO-1_00-1_11_2_5'