HAL_OMAP3: Add GPIO device implementation
Update to API 1.0, and corresponding set of register wide accessor functions.
The tables of permitted pins come from staring at the schematics for any connectors, minus those that are "owned" by the OS.
In the absence of any Touchbook schematics, no GPIO device will be registered for this (given the physical size of the Touchbook it's unlikely there are any usefully accessible headers anyway).

Tested on a BB-xM, toggling the two user LEDs and scanning the 'USER' push button.

Version 1.18. Tagged as 'OMAP3-1_18'