HAL_OMAP4: Add GPIO device implementation
Update to API 1.0, and corresponding set of register wide accessor functions.
The tables of permitted pins come from staring at the schematics for any connectors, minus those that are "owned" by the OS.

Tested on a Pandaboard, toggling the two user LEDs and scanning the 'USER' push button (which, by the way, needs the pullup enabling as there isn't a discrete one).

Version 0.58. Tagged as 'OMAP4-0_58'