ScrnSetup: Improve fading logic when unselecting Native
When the previously configured mode was Native, in order to avoid accessing the colour flags array with a -ve index, the depth was arbitrarily set to 1bpp.
However, if the graphics controller doesn't support 1bpp the search would exit with no matches the the colour/frame rate popups remained greyed and unselectable.
Instead let the loop run up to the highest colour depth (16M), the best match logic will pick the best colour/frame rate combination, but more importantly the GUI controls will be activated.

Version 0.42. Tagged as 'ScrnSetup-0_42'