ThemeSetup: Fix for bogus font menu text if Wimp$Font is unset
If the configured desktop font is 0 (use Wimp$Font) but that variable is unset, such as having installed !Boot afresh having previously used a non-ROM font, the call to getenv() would return a NULL pointer which was copied into the font display field.
Check the result before doing a strcpy().

Version 0.05. Tagged as 'ThemeSetup-0_05'