ZeroPain: Enter logging only role
With the advent of the kernel's compatibility page in Kernel-5_87 and later, much of the utility of ZeroPain allowing buggy apps to plod on is redundant.
However, developers still find its logging useful, since it's easier to read than letting an abort happen (with the compatibility page turned off).
Move the kill switch to 5.26 and remove the emulation of kernel workspace values, leaving two states
a) you have the compatibility page for now, and silence for reads
b) you have ZeroPain loaded, and get logs for reads
and all writes in any mode aborts.

Version 0.08. Tagged as 'ZeroPain-0_08'