HAL_Tungsten: Change progress log append implementation
Use advanced technology including OPENOUT and OPENUP to create and append the progress log, instead of using OSCLI of *If/*Do/*Echo with redirection.
This (by inference) caused BASIC to be kicked out of the application slot when the command was implemented as a disc based application.
Ref https://www.riscosopen.org/forum/forums/11/topics/11472

Version 0.63. Tagged as 'Tungsten-0_63'