InetSetup: Fix for lack of address when only Access is enabled
When only Access is enabled in InetSetup it doesn't work, this is because in Freeway 0.29 (August 1999) a new configuration command FreewayAutoAddress was added, but whose default is off (due to CMOS wiping to zero).
As long as Internet is enabled that's OK because the interface will use whatever address method is chosen for TCP/IP. When only Access is enabled no address is set. In truth table form:
  Access  TCP/IP
  0       X      => Don't care, Access is disabled
  1       0      => need an auto address
  1       1      => use the same address as TCP/IP
Update InetSetup to manage the CMOS bit per the above.

Messages:RMEnsure of Freeway changed to 0.36 (the version when the auto assigned address was moved from 1.x.y.z to 10.x.y.z range).
Save.c:Set or clear the FreewayAutoAddress bit as required.

Version 0.60. Tagged as 'InetSetup-0_60'