ABRelease: Auto generate packages of some disc components
At the end of a DiscDev build, turn the output into packages for transfer outside of the autobuilder.
Which things to package is driven by a package database ('PackageDB') which is cross referenced with the ModuleDB. It also collects licence terms from the respective source directory.
Package suffixes start at '1' and increment if the contents change - ie. it's not pegged off soley the VersionNum, so if a library or tool gets a fix all impacted components will get a new suffix.

DiscDev/clean: Wipe away any intermediate packages
DiscDev/release_autobuild: Create the packages
Makefile: Correct a typo

In this initial stage, only the Diversions are covered. Baby steps...

Version 0.92. Tagged as 'ABRelease-0_92'