ABRelease: Add another bundle to the package list
Also extend the PkgHardDisc4 tool to allow a package name to be different to the component name listed in ModuleDB.
Fix bug when a new package is added to the list (so it's not in the previous hashes) where it would reuse the previous entry's ZIP suffix.

Version 0.95. Tagged as 'ABRelease-0_95'