BuildSys: Add NetUtils to ModuleDB and PlingSystem
Rename dummy module NetUtils2 to NetUtils2. Use NetUtils for the real thing.
RISC OS 3.50 and later had a new enough version of NetFS to not need patching (>=5.79), so on the face of it the simplest solution would be to load the newer NetFS on earlier OS versions. However, that wouldn't work if you were booting via NetFS (compared with loading NetUtils, which can be done while NetFS is being used).

Version 7.35. Tagged as 'BuildSys-7_35'